Split file configuration tests in multiple files
